Call parse_links from django middleware instead of from templates
- Move parse_links into core/utils
- Make parse_links in v1/__init__ a dummy function for now
- Create new middleware class specifically to process external links
